refactor: applying PR commits.
