+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/report/componant/animatedMeter/CaloriesMeter.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.graphics.drawscope.Stroke
 import androidx.compose.ui.unit.dp
 import com.cairosquad.evolvefit.design_system.theme.Theme
+import com.cairosquad.evolvefit.ui.util.toFormattedString
 import org.jetbrains.compose.ui.tooling.preview.Preview
 
 @Composable
@@ -139,12 +140,12 @@ fun CaloriesMeter(
             verticalArrangement = Arrangement.spacedBy(4.dp)
         ) {
             Text(
-                text = takenCalories.toString(),
+                text = takenCalories.toUInt().toFormattedString(),
                 style = Theme.textStyle.headline.largeBold18,
                 color = Theme.color.surfaces.onSurfaceContainer
             )
             Text(
-                text = "From $expectedCalories kcal",
+                text = "From ${expectedCalories.toUInt().toFormattedString()} kcal",
                 style = Theme.textStyle.label.smallRegular12,
                 color = Theme.color.surfaces.onSurfaceVariant
             )
