+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/report/componant/cards/HistoryWorkoutCard.kt
@@ -15,6 +15,9 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
 import com.cairosquad.evolvefit.design_system.theme.Theme
 import com.cairosquad.evolvefit.viewmodel.report.ReportScreenState
+import evolvefit.composeapp.generated.resources.Res
+import evolvefit.composeapp.generated.resources.workout_history
+import org.jetbrains.compose.resources.stringResource
 
 @Composable
 fun HistoryWorkoutCard(
@@ -24,7 +27,7 @@ fun HistoryWorkoutCard(
 ) {
     ReportCard(
         modifier = modifier,
-        title = "Workout History",
+        title = stringResource(Res.string.workout_history),
         value = ""
     ) {
         val workouts = state.workoutHistory.take(3)
